feat(defaultOptions): set scope to false by default
scope conceptually is a good idea, but has too many variations. I want to find a good way to support it, potentially by capping the depth. This still runs into issues with other plugins.

This can still be turned on via a flag. Please open an issue if you wish to have this back by default.